Version 0.2.1 - Use shared observer
(1) Use a shared observer between modules that observe the page.
(2) Keep track of the numbers of post filtered and deleted and display
on page.
(3) fix mispelling in description of extension in manifest.
(4) On adding a user to the filter, clear the textbox with the user name
being added.
(5) Add another bot to the default filter.